refactor: get the SimpleExchange working: can trade; bookOrders show
This is a merge of work by @katelynsills and myself. The work was
interleaved enough that attribution of particular parts is difficult,
so it's all lumped together to make it possible to review.

subscribe to simpleExchange notifier for bookOrders changes
drop extraneous notification code in handler.js
Add back instance reg key and asset and price brand reg keys (#10)
map brands; drop inviteHandle, changed and state from orders

The brandRegKeys are coming out as 'Asset', and 'Price'. We deal with
  that rather than fixing it.