refactor: finish notification from book orders; drop extraneous code
There appear to be two or three old mechanisms for getting updates. I
removed those, and hooked up the new notifier-based version to where
it looked like it should go.

This doesn't work yet, but the brokenness around purses in the UI is
making it hard to see what needs to be fixed next.